Rename map move command module
