+++ commands/cmdsets/economy_map.py
@@ -3,7 +3,7 @@
 from evennia import CmdSet
 from commands.player.cmd_store import CmdStore
 from commands.player.cmd_pokestore import CmdPokestore
-from commands.player.cmdmapmove import CmdMapMove
+from commands.player.cmd_map_move import CmdMapMove
 from commands.player.cmdstartmap import CmdStartMap
