+++ commands/player/cmd_map_move.py
@@ -1,3 +1,5 @@
+"""Command allowing movement within the current map."""
+
 from evennia import Command
